♻️ refactor: applications 파일 통합
1. API 엔드포인트 및 View 구조 개선
 - 지원서 생성와 지원자 목록 조회 API의 엔드포인트 통합
 - 분리되어 있던 View들을 ApplicationAPIView 하나로 통합.
2. 권한 로직 최적화
 - IsRecruitmentAuthor 권한 클래스에서 공고 객체 조회 시, 해당 객체를 request에 저장하도록 하여 View에서 발생하는 중복 DB 조회를 방지하고 성능을 개선
3. 코드 구조 리팩토링 및 정리
 - 여러 개로 분리되어 있던 managers를 하나로 통합, 리팩토링 과정에서 불필요해진 view, manager, service 파일을 삭제
4. 테스트 및 타입 에러 수정
 - 리팩토링으로 인해 변경된 URL 이름, 권한 로직, 누락됨 import등 테스트 코드와 소스 코드에 반영하여 mypy 타입 에러와 테스트 실패 해결.

관련 이슈: #134

# Conflicts:
#	apps/recruitments/urls/recruitments_urls.py

# Conflicts:
#	apps/recruitments/urls/recruitments_urls.py
