+++ apps/applications/managers/application_managers.py
@@ -1,8 +1,9 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any
+from uuid import UUID
 
-from django.db import IntegrityError, models
+from django.db import models
 
 from apps.recruitments.models import Recruitment
 from apps.users.models import User
@@ -12,6 +13,9 @@ if TYPE_CHECKING:  # pragma: no cover
 
 
 class ApplicationManager(models.Manager["Application"]):
+    def get_applications_for_recruitment(self, recruitment_uuid: UUID) -> models.QuerySet["Application"]:
+        """특정 공고에 대한 지원서 목록을 반환하는 매니저 메소드"""
+        return self.filter(recruitment__uuid=recruitment_uuid).select_related("user")
 
     def has_applied(self, user: User, recruitment: Recruitment) -> bool:
         """
