+++ apps/recruitments/urls/recruitments_urls.py
@@ -8,7 +8,7 @@ from apps.recruitments.views.recruitment_views import MyRecruitmentView, Recruit
 urlpatterns = [
     path("", RecruitmentView.as_view(), name="recruitment-list"),
     path("/me", MyRecruitmentView.as_view(), name="recruitment-mylist"),
-    path("/<uuid:recruitment_uuid>/applications", include("apps.applications.urls")),
+    path("/<uuid:recruitment_uuid>/applications", include("apps.applications.urls.recruitments_uuid_urls")),
     path("/<uuid:recruitment_uuid>", RecruitmentDetailView.as_view(), name="recruitment-detail"),
     path("/attachments", RecruitmentFileView.as_view(), name="recruitment-attachments-upload"),
     path("/images", RecruitmentImageView.as_view(), name="recruitment-images-upload"),
